+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet/xlsx_row/xlsx_cell/formula.rb
@@ -28,5 +28,10 @@ module OoxmlParser
       @value = node.text unless node.text.empty?
       self
     end
+
+    # @return [True, False] check if formula empty
+    def empty?
+      !(reference || string_index || type || value)
+    end
   end
 end