+++ spec/workbook/all_formula_values_spec.rb
@@ -10,4 +10,9 @@ describe 'Workbook#all_formula_values' do
     xlsx = OoxmlParser::XlsxParser.parse_xlsx('spec/workbook/all_formula_values_spec/all_formulas_values.xlsx')
     expect(xlsx.all_formula_values.length).to eq(12)
   end
+
+  it 'all_formulas_values should not return values of empty formulas' do
+    xlsx = OoxmlParser::XlsxParser.parse_xlsx('spec/workbook/all_formula_values_spec/empty_formulas.xlsx')
+    expect(xlsx.all_formula_values).to eq(['1.0'])
+  end
 end